requirements.txt
a2wsgi==1.10.8
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioice==0.10.1
aiortc==1.11.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
asgiref==3.8.1
attrs==25.3.0
av==14.3.0
blinker==1.9.0
cachetools==5.5.2
cartesia==1.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
cryptography==44.0.2
daily-python==0.17.0
dataclasses-json==0.6.7
deepgram-sdk==3.8.0
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
fastapi==0.115.12
flask==3.1.0
flatbuffers==25.2.10
flexcache==0.3
flexparser==0.4
frozenlist==1.6.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-speech==2.31.1
google-cloud-texttospeech==2.25.1
google-crc32c==1.7.1
google-genai==1.7.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
iterators==0.2.0
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.9.0
llvmlite==0.44.0
loguru==0.7.3
markdown==3.8
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.1.0
numba==0.61.2
numpy==1.26.4
onnxruntime==1.20.1
openai==1.70.0
packaging==25.0
pillow==11.1.0
pint==0.24.4
pipecat==0.3.0
pipecat-ai==0.0.65
pipecat-ai-flows==0.0.16
pipecatcloud==0.1.5
platformdirs==4.3.7
prompt-toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydub==0.25.1
pyee==13.0.0
pygments==2.19.1
pylibsrtp==0.12.0
pyloudnorm==0.1.1
pyopenssl==25.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
questionary==2.1.0
requests==2.32.3
resampy==0.4.3
rfc3986==1.5.0
rich==14.0.0
rsa==4.9.1
scipy==1.15.2
shellingham==1.5.4
sigtools==4.0.1
simli-ai==0.1.11
six==1.17.0
sniffio==1.3.1
soxr==0.5.0.post1
starlette==0.46.2
sympy==1.13.3
synchronicity==0.7.7
toml==0.10.2
tqdm==4.67.1
typer==0.13.1
types-python-dateutil==2.9.0.20241206
typing-extensions==4.13.2
typing-inspect==0.9.0
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.32.1
wcwidth==0.2.13
websockets==13.1
werkzeug==3.1.3
yarl==1.20.0
